chromium/third_party/blink/web_tests/virtual/out-of-blink-cspee/external/wpt/content-security-policy/embedded-enforcement/subsumption_algorithm-hashes-expected.txt

This is a testharness.js-based test.
PASS 'sha256-abc123' is properly subsumed.
FAIL Returned should not include hashes not present in required csp. assert_equals: expected (undefined) undefined but got (boolean) true
PASS 'sha256-abc123' is properly subsumed with other sources.
PASS Hashes do not have to be present in returned csp.
FAIL Hashes do not have to be present in returned csp but must not allow all inline behavior. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Other expressions have to be subsumed. assert_equals: expected (undefined) undefined but got (boolean) true
PASS Other expressions have to be subsumed but 'unsafe-inline' gets ignored.
PASS Effective policy is properly found.
FAIL Required csp must allow 'sha256-abc123'. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Effective policy is properly found where 'sha256-abc123' is not subsumed. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL 'sha256-abc123' is not subsumed by 'sha256-abc456'. assert_unreached: No message should be sent from the frame. Reached unreachable code
PASS Effective policy now does not allow 'sha256-abc123'.
PASS Effective policy is properly found where 'sha256-abc123' is not part of it.
Harness: the test ran to completion.
